Feature: Cloudinary react

+++ routes/meetup.routes.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const mongoose = require("mongoose");
+const fileUploader = require("../config/cloudinary.config");
 
 const Meetup = require("../models/Meetup.model");
 const User = require("../models/User.model");
@@ -10,6 +11,20 @@ router.get("/", (req, res, next) => {
   Meetup.find()
     .then((allMeetup) => res.json(allMeetup))
     .catch((error) => res.json(error));
+});
+
+// POST Route that receives the image, sends it to Cloudinary via the fileUploader and returns the image URL
+router.post("/upload", fileUploader.single("eventImage"), (req, res, next) => {
+  console.log("file is: ", req.file);
+
+  if (!req.file) {
+    next(new Error("No image uploaded!"));
+    return;
+  }
+  // Get the URL of the uploaded file and send it as a response.
+  // 'fileUrl' can be any name, just make sure you remember to use the same when accessing it on the frontend
+
+  res.json({ fileUrl: req.file.path });
 });
 
 // POST /meetup --> create a meetup
@@ -25,7 +40,7 @@ router.post("/create", (req, res, next) => {
     eventImage,
     eventDateAndTime,
     attendees,
-    author
+    author,
   } = req.body;
 
   Meetup.create({
@@ -39,7 +54,7 @@ router.post("/create", (req, res, next) => {
     eventImage,
     eventDateAndTime,
     attendees,
-    author
+    author,
   })
     .then((response) => res.json(response))
     .catch((error) => res.json(error));
@@ -73,9 +88,14 @@ router.put("/edit/:meetupId", (req, res, next) => {
     .catch((error) => res.json(error));
 });
 
-// DELETE /meetup/edit/6419f68980b77e9438d2e48c --> delete meetup
+// DELETE /meetup/edit/:meetupId --> delete meetup
 router.delete("/edit/:meetupId", (req, res, next) => {
   const { meetupId } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(meetupId)) {
+    res.status(400).json({ message: "Specified id is not valid" });
+    return;
+  }
 
   Meetup.findByIdAndDelete(meetupId)
     .then((deleteMeetup) => res.json(deleteMeetup))
